pyproject.toml
[project]
name = "chat2llms"
version = "0.0.4"
description = "与大模型交互的RAG问答系统"
readme = "README.md"
maintainers = [{ name = "Jacob Zheng", email = "1453366421@qq.com" }]
requires-python = ">=3.8"
dependencies = [
    "haystack-ai==2.4.0",
    "fastapi==0.112.2",
    "uvicorn==0.30.6",
    "python-dotenv==1.0.0",
    "sentence-transformers>=2.3.0",
    "qdrant-haystack==4.2.0",
    "ollama-haystack==0.0.7",
]

[project.optional-dependencies]
dev = []

[build-system]
requires = ["setuptools>=61.0.0"]
build-backend = "setuptools.build_meta"


[tool.setuptools]
# packages = ["backend","backend.llms","backend.embedding"]